--- apps/site/src/app/(main)/apply/sections/Form/HackerForm.module.scss
@@ -1,20 +1,0 @@
-.input {
-	background-color: #e1e1e1;
-	color: #000;
-	height: 40px;
-	padding: 5px;
-	border-radius: 5px;
-	font-size: 18px;
-}
-
-.label {
-	font-size: 18px;
-	margin-bottom: 8px;
-}
-
-.upload {
-	& > * {
-		// prevents child elements from interfering with drag & drop
-		pointer-events: none;
-	}
-}
