+++ apps/site/src/app/(main)/apply/sections/Form/HackerForm.tsx
@@ -1,10 +1,11 @@
 import AgeInformation from "@/lib/components/forms/shared/AgeInformation";
 import Form from "@/lib/components/forms/shared/Form";
 import SchoolInformation from "@/lib/components/forms/shared/SchoolInformation";
+import ResumeInformation from "@/lib/components/forms/shared/ResumeInformation";
 
 import BasicInformation from "./HackerBasicInformation";
 import ProfileInformation from "./ProfileInformation";
-import ResumeInformation from "./ResumeInformation";
+
 
 export default function HackerForm() {
 	return (
